+++ tests/p1/tracking/not-implemented.spec.ts
@@ -1,0 +1,8 @@
+import { test } from '@playwright/test';
+
+test('Not implemented', () => {
+  throw new Error(
+    'This tracking tag type is recognised but not yet implemented. ' +
+    'Raise a request to add the spec for this tag type.'
+  );
+});